#2180 Moved DungeonRoute relationship classes to its own namespace.
