+++ app/Http/Controllers/Ajax/AjaxDungeonRouteController.php
@@ -28,8 +28,8 @@ use App\Logic\MapContext\DungeonRouteProperties;
 use App\Logic\MDT\Exception\ImportWarning;
 use App\Models\Dungeon;
 use App\Models\DungeonRoute;
-use App\Models\DungeonRouteFavorite;
-use App\Models\DungeonRouteRating;
+use App\Models\DungeonRoute\DungeonRouteFavorite;
+use App\Models\DungeonRoute\DungeonRouteRating;
 use App\Models\Expansion;
 use App\Models\GameServerRegion;
 use App\Models\PublishedState;
