+++ app/Models/DungeonRoute.php
@@ -6,6 +6,14 @@ use App\Http\Requests\DungeonRoute\DungeonRouteTemporaryFormRequest;
 use App\Logic\Structs\LatLng;
 use App\Models\AffixGroup\AffixGroup;
 use App\Models\CombatLog\ChallengeModeRun;
+use App\Models\DungeonRoute\DungeonRouteAffixGroup;
+use App\Models\DungeonRoute\DungeonRouteAttribute;
+use App\Models\DungeonRoute\DungeonRouteEnemyRaidMarker;
+use App\Models\DungeonRoute\DungeonRouteFavorite;
+use App\Models\DungeonRoute\DungeonRoutePlayerClass;
+use App\Models\DungeonRoute\DungeonRoutePlayerRace;
+use App\Models\DungeonRoute\DungeonRoutePlayerSpecialization;
+use App\Models\DungeonRoute\DungeonRouteRating;
 use App\Models\Enemies\OverpulledEnemy;
 use App\Models\Enemies\PridefulEnemy;
 use App\Models\Floor\Floor;
