+++ app/Models/DungeonRoute/DungeonRoutePlayerSpecialization.php
@@ -1,7 +1,9 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\DungeonRoute;
 
+use App\Models\CharacterClassSpecialization;
+use App\Models\DungeonRoute;
 use Eloquent;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
