+++ app/Models/DungeonRoute/DungeonRouteThumbnailJob.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models\DungeonRoute;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ *
+ */
+class DungeonRouteThumbnailJob extends Model
+{
+
+}
