+++ app/SeederHelpers/RelationImport/Parsers/Relation/DungeonRouteEnemyRaidMarkersRelationParser.php
@@ -3,7 +3,7 @@
 namespace App\SeederHelpers\RelationImport\Parsers\Relation;
 
 use App\Models\DungeonRoute;
-use App\Models\DungeonRouteEnemyRaidMarker;
+use App\Models\DungeonRoute\DungeonRouteEnemyRaidMarker;
 
 class DungeonRouteEnemyRaidMarkersRelationParser implements RelationParserInterface
 {
