+++ app/SeederHelpers/RelationImport/Parsers/Relation/DungeonRoutePlayerClassRelationParser.php
@@ -3,7 +3,7 @@
 namespace App\SeederHelpers\RelationImport\Parsers\Relation;
 
 use App\Models\DungeonRoute;
-use App\Models\DungeonRoutePlayerClass;
+use App\Models\DungeonRoute\DungeonRoutePlayerClass;
 
 class DungeonRoutePlayerClassRelationParser implements RelationParserInterface
 {
