+++ app/SeederHelpers/RelationImport/Parsers/Relation/DungeonRoutePlayerSpecializationRelationParser.php
@@ -3,7 +3,7 @@
 namespace App\SeederHelpers\RelationImport\Parsers\Relation;
 
 use App\Models\DungeonRoute;
-use App\Models\DungeonRoutePlayerSpecialization;
+use App\Models\DungeonRoute\DungeonRoutePlayerSpecialization;
 
 class DungeonRoutePlayerSpecializationRelationParser implements RelationParserInterface
 {
