+++ app/Service/CombatLog/Builders/CreateRouteBodyDungeonRouteBuilder.php
@@ -6,7 +6,7 @@ use App;
 use App\Models\AffixGroup\AffixGroup;
 use App\Models\Dungeon;
 use App\Models\DungeonRoute;
-use App\Models\DungeonRouteAffixGroup;
+use App\Models\DungeonRoute\DungeonRouteAffixGroup;
 use App\Models\Faction;
 use App\Models\Floor\Floor;
 use App\Models\PublishedState;
